requirements.txt
toml==0.10.2
pyyaml==6.0.1
requests==2.32.3
pytest==8.3.2
